Refactor: Input, Calendar 파일 위치 변경, Home폴더 home으로 변경
