+++ pages/test.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import styled from 'styled-components';
-import Input from '@/components/common/Input';
+import Input from '@/components/common/Input/Input';
 
 /**
  * 컴포넌트 실험용 페이지입니다. (컴포넌트 구현 완료 후 삭제 예정)
